filehandler refactoring
